QOL improvements (read description)
Okay so, I added a new extension function called "SeekL", this would allow us to put custom code when doing a seek, so that if an error pops out we can do something other than just make the program crash.
I tried to implement NLog, but frankly I don't fully know how it works.
Also I made a new little custom icon for the project, and I improved the title bar to show a warning and the file name that is being edited as well.